Moved generic utilities at adapters to the engine
