+++ klab.engine/pom.xml
@@ -858,6 +858,36 @@
 			<artifactId>commons-exec</artifactId>
 			<version>1.2</version>
 		</dependency>
+	
+		<!-- https://mvnrepository.com/artifact/org.jetbrains.kotlin/kotlin-stdlib -->
+		<dependency>
+				<groupId>org.jetbrains.kotlin</groupId>
+				<artifactId>kotlin-stdlib</artifactId>
+				<version>1.9.0</version>
+		</dependency>
+	
+		<!-- https://mvnrepository.com/artifact/io.minio/minio -->
+		<dependency>
+				<groupId>io.minio</groupId>
+				<artifactId>minio</artifactId>
+				<version>8.5.4</version>
+				<exclusions>
+					<exclusion>
+						<groupId>com.google.guava</groupId>
+						<artifactId>guava</artifactId>
+					</exclusion>
+					<exclusion>
+						<groupId>org.jetbrains.kotlin</groupId>
+						<artifactId>kotlin-stdlib</artifactId>
+					</exclusion>
+				</exclusions>
+			</dependency>
+			<dependency>
+				<groupId>org.junit.jupiter</groupId>
+				<artifactId>junit-jupiter-api</artifactId>
+				<version>5.9.3</version>
+				<scope>test</scope>
+			</dependency>
 
 	</dependencies>
 </project>
