+++ klab.engine/src/main/java/org/integratedmodelling/klab/utils/s3/S3URLUtils.java
@@ -1,4 +1,4 @@
-package org.integratedmodelling.klab;
+package org.integratedmodelling.klab.utils.s3;
 
 public class S3URLUtils {
     final public static String AWS_ENDPOINT = "https://s3.amazonaws.com";
